NuttX: update ADC block to get the number of configured channels
This commit updates the ADC block to get the number of configured channels
directly from the driver by using ioctl command ANIOC_GET_NCHANNELS. This
replace the current state when the number of configured channels had to be
defined in the code by the user. Assigning the values read by ADC to the
block outputs is also change so the number of channels read by the block
does not have to be equal to the number of configured channels. This
allows the block to be more flexible for different configurations.

This commit also change the name of the files from nuttx_Adc to nuttx_ADC
as is the standard for the other blocks.
