--- resources/blocks/rcpBlk/NuttX/nuttx_AdcBlk.py
@@ -1,22 +1,0 @@
-
-from supsisim.RCPblk import RCPblk
-from scipy import size
-
-def nuttx_AdcBlk(pout, devname, ch, res, umin, umax):
-    """
-
-    Call:   nuttx_AdcBlk(pout, devname)
-
-    Parameters
-    ----------
-       pout: connected output port(s)
-       devname : Port
-
-    Returns
-    -------
-       blk: RCPblk
-
-    """
-
-    blk = RCPblk('nuttx_Adc', [], pout, [0,0], 0, [umin, umax], [ch-1, res, 0], devname)
-    return blk
